fix(drift-check): address Copilot review comments
- apiDriftCheck.ts: restore `params: ["chainId"]` on the three routing
  wrapper base-endpoint contracts (/config/hub/assets, /config/swap/tokens,
  /config/money-market/tokens). With `params: []` the `allowToolExtra`
  field was dead code because the diff never saw `chainId` as extra.
  Now the contract honestly reflects what the tool declares, and
  `allowToolExtra` does its job.

- sodaxApi.ts getUserTransactions: switch `if (options?.offset)` and
  `if (options?.limit)` to `!== undefined` so `offset: 0` (the valid
  first-page offset) is actually sent to upstream instead of silently
  dropped by the truthy check.

- sodaxApi.ts getOrderbook: tighten `offset?: number` to `offset: number`
  to match the tool's required schema. The function is only called from
  the tool, so this keeps the type contract consistent top-to-bottom.

- tools/sodaxApi.ts sodax_get_partners.chainId: tighten `z.number()` to
  `z.number().int()` — the upstream requires an integer (400s otherwise),
  so failing at the Zod layer gives a cleaner error.

+++ src/tools/sodaxApi.ts
@@ -406,7 +406,7 @@ export function registerSodaxApiTools(server: McpServer): void {
     "sodax_get_partners",
     "List all SODAX integration partners including wallets, DEXs, and other protocols",
     {
-      chainId: z.number().optional()
+      chainId: z.number().int().optional()
         .describe("Filter partners by numeric chain ID (e.g. 146 for Sonic)"),
       format: z.nativeEnum(ResponseFormat).optional().default(ResponseFormat.MARKDOWN)
         .describe("Response format: 'json' for raw data or 'markdown' for formatted text")
